requirements.txt
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
coverage==7.10.7
cssselect==1.3.0
cssutils==2.11.1
flake8==7.3.0
idna==3.10
iniconfig==2.1.0
Jinja2==3.1.6
lxml==6.0.2
MarkupSafe==3.0.3
mccabe==0.7.0
more-itertools==10.8.0
packaging==25.0
pluggy==1.6.0
premailer==3.10.0
pycodestyle==2.14.0
pyflakes==3.4.0
Pygments==2.19.2
pytest==8.4.2
pytest-cov==7.0.0
pytest-mock==3.15.1
python-dotenv==1.1.1
requests==2.32.5
urllib3==2.5.0
